Moved file editing and line/cell replacement to sector model
